Add SharedLogic with common methods for exchanges
