+++ app/logic/sol_prices/coin_gecko_logic.rb
@@ -1,7 +1,7 @@
 module SolPrices::CoinGeckoLogic
   include SolPrices::Parsers::CoinGecko
 
-  def get_prices_from_days
+  def get_ohlc_prices
     lambda do |p|
       response = p.payload[:client].ohlc(days: p.payload[:days])
       prices = prices_from_ohlc(response)
@@ -52,26 +52,6 @@ module SolPrices::CoinGeckoLogic
     end
   end
 
-  def add_epoch
-    lambda do |p|
-      # We will run this pipeline just after the midnight 
-      # so we can assume that last Epoch History will be correct.
-      # 
-      # We can later find the closest record to the date to be more specific.
-      epoch_testnet = EpochHistory.where(network: 'testnet').last
-      epoch_mainnet = EpochHistory.where(network: 'mainnet').last
-
-      Pipeline.new(
-        200, p.payload.merge(
-          epoch_testnet: epoch_testnet&.epoch,
-          epoch_mainnet: epoch_mainnet&.epoch
-        )
-      )
-    rescue StandardError => e
-      Pipeline.new(500, p.payload, 'Error from add_epoch', e)
-    end
-  end
-
   # For one, yesterday price.
   def save_sol_price
     lambda do |p|
@@ -82,30 +62,13 @@ module SolPrices::CoinGeckoLogic
       datetime_from_exchange = p.payload.dig(:sol_price, :datetime_from_exchange)
 
       SolPrice.where(
-        exchange: SolPrice.exchanges[:coin_gecko],
+        exchange: p.payload[:exchange],
         datetime_from_exchange: datetime_from_exchange
       ).first_or_create(p.payload[:sol_price])
               
       Pipeline.new(200, p.payload)
     rescue StandardError => e
       Pipeline.new(500, p.payload, 'Error from save_sol_price', e)
-    end
-  end
-
-  # For batch with prices.
-  def save_sol_prices
-    lambda do |p|
-      p.payload[:prices_from_exchange].each do |sol_price|
-        datetime_from_exchange = sol_price[:datetime_from_exchange]
-        SolPrice.where(
-          exchange: SolPrice.exchanges[:coingecko],
-          datetime_from_exchange: datetime_from_exchange
-        ).first_or_create(sol_price)
-      end
-              
-      Pipeline.new(200, p.payload)
-    rescue StandardError => e
-      Pipeline.new(500, p.payload, 'Error from save_sol_prices', e)
     end
   end
 end
