+++ app/models/sol_price.rb
@@ -17,6 +17,6 @@
 #  updated_at             :datetime         not null
 #
 class SolPrice < ApplicationRecord
-  enum exchange: { coingecko: 0, ftx: 1 }, _prefix: true
+  enum exchange: { coin_gecko: 0, ftx: 1 }, _prefix: true
   enum currency: { usd: 0 }, _prefix: true
 end
