+++ script/sol_prices/ftx/gather_yesterday_prices.rb
@@ -1,37 +1,15 @@
 # frozen_string_literal: true
 
-# Argument is optional, when is omitted max will be used.
-# Valid values: 1/7/14/30/90/180/365/'max'. Defaults to 7 days.
-# 
-# RAILS_ENV=production bundle exec rails r script/gather_historical_prices_from_coin_gecko.rb 365 
-
-require_relative '../../../config/environment'
-require 'coin_gecko_logic'
-
-include CoinGeckoLogic
-
-# Create our initial payload with the input values
-initial_payload = {
-  client: SolPrices::ApiWrappers::CoinGecko.new,
-  datetime: Time.now.utc.beginning_of_day.to_datetime,
-  days: ARGV[0] || 'max'
-}
-
-p = Pipeline.new(200, initial_payload)
-            .then(&get_prices_from_days)
-            .then(&save_sol_prices)
-
-# frozen_string_literal: true
-
 # RAILS_ENV=production bundle exec rails r script/sol_prices/ftx/gather_yesterday_prices_from.rb 
 
 require_relative '../../../config/environment'
-require 'ftx_logic'
 
-include FtxLogic
+include SolPrices::FtxLogic
+include SolPrices::SharedLogic
 
 # Create our initial payload with the input values
 initial_payload = {
+  exchange: SolPrice.exchanges[:ftx],
   client: SolPrices::ApiWrappers::FtxMarket.new,
   start_time: Date.yesterday.to_datetime,
   end_time: Date.yesterday.to_datetime,
@@ -39,6 +17,6 @@ initial_payload = {
 }
 
 p = Pipeline.new(200, initial_payload)
-            .then(&get_prices_from_days)
+            .then(&get_historical_prices)
             .then(&save_sol_prices)
