+++ test/vcr_cassettes/logic/sol_prices/shared/test__save_sol_prices_coin_gecko.yml
