+++ test/vcr_cassettes/logic/sol_prices/shared/test__save_sol_prices_ftx.yml
@@ -1,0 +1,68 @@
+---
+http_interactions:
+- request:
+    method: get
+    uri: https://ftx.com/api/markets/SOL/USD/candles?resolution=86400&start_time=1631059200
+    body:
+      encoding: US-ASCII
+      string: ''
+    headers:
+      Accept-Encoding:
+      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
+      Accept:
+      - "*/*"
+      User-Agent:
+      - Ruby
+      Host:
+      - ftx.com
+      Ftx-Key:
+      - AUk9zEtHT7Jan9XINgEBpMOzp-XP7F28qExTYYj1
+      Ftx-Sign:
+      - a3133ef221e0d2b87f72efeb0ba5481d2e8305e32d25f0dbf74cec80ba1f41b2
+      Ftx-Ts:
+      - '1631283786000'
+  response:
+    status:
+      code: 200
+      message: OK
+    headers:
+      Date:
+      - Fri, 10 Sep 2021 14:23:07 GMT
+      Content-Type:
+      - application/json
+      Transfer-Encoding:
+      - chunked
+      Connection:
+      - keep-alive
+      Vary:
+      - Accept-Encoding
+      - Accept-Encoding
+      - Origin
+      Cache-Control:
+      - public, max-age=60
+      Etag:
+      - W/"2c19cc116dddcfba377cf0ebe03af54d"
+      X-Cache-Status:
+      - MISS
+      Cf-Cache-Status:
+      - EXPIRED
+      Expect-Ct:
+      - max-age=604800, report-uri="https://report-uri.cloudflare.com/cdn-cgi/beacon/expect-ct"
+      Strict-Transport-Security:
+      - max-age=31536000; includeSubDomains; preload
+      X-Content-Type-Options:
+      - nosniff
+      Server:
+      - cloudflare
+      Cf-Ray:
+      - 68c94ff4bd32f290-WAW
+      Alt-Svc:
+      - h3=":443"; ma=86400, h3-29=":443"; ma=86400, h3-28=":443"; ma=86400, h3-27=":443";
+        ma=86400
+    body:
+      encoding: ASCII-8BIT
+      string: '{"success":true,"result":[{"startTime":"2021-09-08T00:00:00+00:00","time":1631059200000.0,"open":173.515,"high":198.2125,"low":146.6825,"close":190.9625,"volume":571066472.04745},{"startTime":"2021-09-09T00:00:00+00:00","time":1631145600000.0,"open":190.9625,"high":216.0,"low":177.85,"close":187.7125,"volume":624323742.34275},{"startTime":"2021-09-10T00:00:00+00:00","time":1631232000000.0,"open":187.7125,"high":197.3675,"low":173.0,"close":175.7125,"volume":208491269.715275}]}
+
+        '
+  recorded_at: Fri, 10 Sep 2021 14:23:07 GMT
+recorded_with: VCR 6.0.0
